[#176782168] Refactor
