+++ api_cgn.yaml
@@ -20,7 +20,7 @@ paths:
         "200":
           description: CGN status details.
           schema:
-            $ref: "#/definitions/CardStatus"
+            $ref: "#/definitions/Card"
         "401":
           description: Bearer token null or expired.
         "404":
@@ -95,18 +95,18 @@ definitions:
     $ref: "https://raw.githubusercontent.com/pagopa/io-functions-commons/v17.3.0/openapi/definitions.yaml#/ProblemJson"
   InstanceId:
     $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/master/openapi/index.yaml#/definitions/InstanceId"
-  CommonCardStatus:
-    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CommonCardStatus"
-  CardPendingStatus:
-    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardPendingStatus"
-  CardActivatedStatus:
-    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardActivatedStatus"
-  CardRevokedStatus:
-    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardRevokedStatus"
-  CardExpiredStatus:
-    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardExpiredStatus"
-  CardStatus:
-    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardStatus"
+  CommonCard:
+    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CommonCard"
+  CardPending:
+    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardPending"
+  CardActivated:
+    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardActivated"
+  CardRevoked:
+    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardRevoked"
+  CardExpired:
+    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CardExpired"
+  Card:
+    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/Card"
   CgnActivationDetail:
     $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn/176782168_add_eyca_async_activation/openapi/index.yaml#/definitions/CgnActivationDetail"
